requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
statsmodels
prophet
scikit-learn
u8darts
convertdate
